Move _impl::width and _impl::set_width into bits/num_traits/
