+++ include/cnl/num_traits.h
@@ -38,22 +38,6 @@ namespace cnl {
     // cnl::is_integral
 
     using _impl::is_integral;
-
-    ////////////////////////////////////////////////////////////////////////////////
-    // cnl::_impl::width / set_width
-
-    namespace _impl {
-        template<typename T>
-        struct width : std::integral_constant<int, digits<T>::value+is_signed<T>::value> {
-        };
-
-        template<typename T, int Bits>
-        struct set_width : set_digits<T, Bits - is_signed<T>::value> {
-        };
-
-        template<typename T, int Bits>
-        using set_width_t = typename set_width<T, Bits>::type;
-    }
 }
 
 #endif  // CNL_NUM_TRAITS
